backend/requirements.txt
fastapi
uvicorn
google-genai
google-generativeai
python-dotenv   
axios